requirements.txt
docopt==0.6.2
numpy==1.23.1
pandas==1.4.3
pydicom==2.3.0
pykwalify==1.8.0
pyradiomics==3.0.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SimpleITK==2.1.1.2
six==1.16.0
